Add navbar and move categories to navbar
