+++ package.json
@@ -8,8 +8,9 @@
     "@testing-library/user-event": "^12.8.3",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react-dom": "^17.0.2",
     "react-redux": "^7.2.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "redux": "^4.0.5",
     "web-vitals": "^1.1.0"
