+++ src/components/Routes.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import App from './App';
+import Navbar from './Navbar';
+
+const Routes = () => (
+  <BrowserRouter>
+    <Navbar />
+    <Switch>
+      <Route path="/" component={App} exact />
+    </Switch>
+  </BrowserRouter>
+);
+
+export default Routes;
