+++ src/containers/BooksList.js
@@ -3,16 +3,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import Book from '../components/Book';
-import { removeBook, changeFilter } from '../actions/index';
-import CategoryFilter from '../components/CategoryFilter';
+import { removeBook } from '../actions/index';
 
 const BooksList = ({
-  books, filter, changeCat, delete: handleRemoveBook,
+  books, filter, delete: handleRemoveBook,
 }) => {
-  const changeCategory = e => {
-    changeCat(e.target.value);
-  };
-
   const filteredBooks = () => {
     if (filter === 'All') {
       return books.sort((a, b) => a.id - b.id);
@@ -22,7 +17,6 @@ const BooksList = ({
 
   return (
     <>
-      <CategoryFilter handleFilter={changeCategory} />
       <table>
         <thead>
           <tr>
@@ -55,14 +49,12 @@ BooksList.propTypes = {
   ).isRequired,
   delete: PropTypes.func.isRequired,
   filter: PropTypes.string.isRequired,
-  changeCat: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = state => ({ books: state.books, filter: state.filter });
 
 const mapDispatchToProps = dispatch => ({
   delete: book => dispatch(removeBook(book)),
-  changeCat: category => dispatch(changeFilter(category)),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(BooksList);
