+++ src/styles/App.css
@@ -1,4 +1,9 @@
-html {
+* {
   padding: 0;
   margin: 0;
+  box-sizing: border-box;
+}
+
+body {
+  background-color: #fafafa;
 }
